feature: Rename the products button
