+++ src/components/ProductSlider/productSlider.css
@@ -95,27 +95,6 @@
   margin-top: 1rem;
 }
 
-.productAddCart {
-  width: 100%;
-  display: block;
-  height: 36px;
-  color: #001738;
-  font-weight: bolder;
-  text-align: center;
-  line-height: 36px;
-  font-size: 14px;
-  text-transform: uppercase;
-  background: #e3e4e6;
-  border-radius: 10px;
-  text-decoration: none;
-  border: 0;
-}
-
-.productAddCart:hover {
-  cursor: pointer;
-  background: #cecfd2;
-}
-
 .noActive {
   background: #e3e4e6;
   color: #cecfd2 !important;
