+++ src/utils/index.tsx
@@ -1,0 +1,5 @@
+export * from './addToCart'
+export * from './getFormatMoney'
+export * from './getMobileOs'
+export * from './getProducts'
+export * from './getRandomColor'
